refactor: move ACL rule parsing out of general into its own core/lib/acl domain
ACL rule parsing/validation is a named buildcage domain concept (like
provenance/ and report/), not cross-cutting infrastructure — general/
should hold only the truly generic ActionError base class and
errorMessage helper.
